add work - first time working :)

+++ src/main/java/org/example/Fsrs.java
@@ -14,7 +14,7 @@ public class Fsrs {
     static final float WEIGHTS[] = new float[] {0.4f, 0.6f, 2.4f, 5.8f, 4.93f, 0.94f, 0.86f, 0.01f, 1.49f, 0.14f, 0.94f, 2.18f, 0.05f,
             0.34f, 1.26f, 0.29f, 2.61f};
 
-    static final float FACTOR = 19 / 81;
+    static final double FACTOR = 19.0 / 81.0;
 
 
 }
